requirements.txt
anthropic
psycopg2-binary
pandas
pypdf
python-dotenv
